refactor(settings): move catalogue/guidance/altitude/site settings out of features/targets to shared (DS-3)
Dissolves the features/settings <-> features/targets two-way dependency edge.

Moved to shared/planner/:
- altitude-settings.ts + test
- catalogue-settings.ts
- guidance-settings.ts + test

Moved to shared/observing-sites/:
- observing-sites/ directory (all 8 files)

Updated import paths in 19 consumers:
features/targets/{TargetsPage,TargetsTable,PlannerComputedFor,site-gate,
useTargetTonight,useTargetsPageFilters,useTargetsTableRows,planner-altitude,
planner-astronomy,planner-derive,table-model,TargetDetailV2} and their tests,
plus features/settings/{PlannerSettings,CatalogueSettingsControl},
features/setup/{SetupWizard,StepSite}, and app/Shell.

Circular dep gate: 0 new cycles.
