+++ apps/desktop/src/features/setup/steps/StepSite.a11y.test.tsx
@@ -12,7 +12,7 @@ import {
   type SiteStepState,
 } from './StepSite';
 
-vi.mock('@/features/targets/observing-sites/SiteLocationPicker', () => ({
+vi.mock('@/shared/observing-sites/SiteLocationPicker', () => ({
   SiteLocationPicker: () => null,
 }));
