+++ apps/desktop/src/features/setup/steps/StepSite.tsx
@@ -16,13 +16,13 @@ import { m } from '@/lib/i18n';
 import {
   localTimezone,
   ianaTimezones,
-} from '@/features/targets/observing-sites/iana-timezones';
-import type { Twilight } from '@/features/targets/observing-sites/observer-site';
+} from '@/shared/observing-sites/iana-timezones';
+import type { Twilight } from '@/shared/observing-sites/observer-site';
 
 // Leaflet is ~240 KB raw. Lazy-load so it splits to its own chunk and does
 // not inflate the setup-wizard bundle parsed on first run.
 const SiteLocationPicker = lazy(() =>
-  import('@/features/targets/observing-sites/SiteLocationPicker').then((m) => ({
+  import('@/shared/observing-sites/SiteLocationPicker').then((m) => ({
     default: m.SiteLocationPicker,
   })),
 );
