+++ apps/desktop/src/features/targets/TargetsPage.tsx
@@ -38,13 +38,16 @@ import { TargetDetail } from './TargetDetail';
 import { useTargets } from './store';
 import { PLANNER_CATALOGS, type CatalogueId } from './planner-catalog';
 import { TargetsTable } from './TargetsTable';
-import { useAltitudeThreshold } from './altitude-settings';
+import { useAltitudeThreshold } from '@/shared/planner/altitude-settings';
 import { useObservingNight } from './astro/observing-night';
 import { computeObservingNight, type ObservingNight } from './astro/moon-state';
 import { useObserverSiteExists } from './site-gate';
 import { PlannerDatePicker } from './PlannerDatePicker';
 import { PlannerComputedFor } from './PlannerComputedFor';
-import { useGuidanceParams, loadGuidanceParams } from './guidance-settings';
+import {
+  useGuidanceParams,
+  loadGuidanceParams,
+} from '@/shared/planner/guidance-settings';
 import { usePlannerSensorConfig } from './planner-sensor';
 import { recommendationLabel } from './FilterBadges';
 import type { Recommendation } from './astro/moon-avoidance';
