+++ apps/desktop/src/features/targets/useTargetsTableRows.ts
@@ -22,7 +22,7 @@ import {
   type RowMoonGeometry,
 } from './astro/row-planning';
 import type { MoonAvoidanceParams } from './astro/moon-avoidance';
-import type { ObserverSite } from './observing-sites/observer-site';
+import type { ObserverSite } from '@/shared/observing-sites/observer-site';
 import {
   buildTargetAccessors,
   flattenGroups,
