+++ apps/desktop/src/shared/observing-sites/ObservingSites.test.tsx
