+++ apps/desktop/src/shared/observing-sites/ObservingSites.tsx
