+++ apps/desktop/src/shared/observing-sites/SiteLocationPicker.test.tsx
